# backend/app/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "GUAM Medical license form",
    "description": "Medical License",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "52.35.218.163:9000",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/register"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "type": "object",
              "title": "Status",
              "properties": {
                "status": {
                  "type": "number",
                  "title": "StatusItem",
                  "example": "200"
                },
                "data": {
                  "type": "object",
                  "title": "StatusItem",
                  "properties": {
                    "user": {
                      "type": "object",
                      "title": "StatusItem",
                      "properties": {
                        "username": {
                          "type": "string",
                          "example": "shah.developer9@gmail.com"
                        },
                        "pool": {
                          "type": "object",
                          "properties": {
                            "userPoolId": {
                              "type": "string",
                              "example": "us-east-1_3zFrmeF6y"
                            },
                            "clientId": {
                              "type": "string",
                              "example": "77s5hu3poe0ti9ljpvhialiqqd"
                            },
                            "client": {
                              "type": "object",
                              "title": "StatusItem",
                              "properties": {
                                "endpoint": {
                                  "type": "string",
                                  "example": "https://cognito-idp.us-east-1.amazonaws.com/"
                                },
                                "userAgent": {
                                  "type": "string",
                                  "example": "aws-amplify/0.1.x js"
                                }
                              }
                              }
                            },
                            "advancedSecurityDataCollectionFlag": {
                              "type": "string",
                              "example": "true"
                            }
                          },
                        "Session": {
                          "type": "string",
                          "example": "fc05d545-6def-459d-8bbc-cf630ea98fb6"
                          },
                        "client": {
                          "type": "object",
                          "properties": {
                            "endpoint": {
                              "type": "string",
                              "example": "https://cognito-idp.us-east-1.amazonaws.com/"
                            },
                            "userAgent": {
                              "type": "string",
                              "example": "aws-amplify/0.1.x js"
                            }
                          }
                        },
                        "signInUserSession": {
                              "type": "string",
                              "example": "null"
                          },
                        "authenticationFlowType": {
                              "type": "string",
                              "example": "USER_SRP_AUTH"
                        },
                        "keyPrefix": {
                              "type": "string",
                              "example": "CognitoIdentityServiceProvider.77s5hu3poe0ti9ljpvhialiqqd"
                        },
                        "userDataKey": {
                              "type": "string",
                              "example": "CognitoIdentityServiceProvider.77s5hu3poe0ti9ljpvhialiqqd.shah.developer9@gmail.com.userData"
                          }
                        }
                      }
                    }
                  }
                },
                "userConfirmed": {
                  "type": "string",
                  "example": "false"
                },
                "userSub": {
                  "type": "string",
                  "example": "cb157073-c54f-455f-87a9-7b53da0798d7"
                }
              }
            }
          }
        }
      },
    "/auth/login": {
      "post": {
        "tags": [
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "type": "object",
              "title": "Status",
              "properties": {
                "status": {
                  "type": "number",
                  "title": "StatusItem",
                  "example": "200"
                },
                "data": {
                  "type": "object",
                  "title": "StatusItem",
                  "properties": {
                    "idToken": {
                      "type": "object",
                      "title": "StatusItem",
                      "properties": {
                        "jwtToken": {
                          "type": "string",
                          "example": "eyJraWQiOiJ3WHpXNGxjMmdkRmY2TmYzd2d0XC8welhIRWx3K1dCUmxxVzZaQXV......"
                        },
                        "payload": {
                          "type": "object",
                          "properties": {
                            "sub": {
                              "type": "string",
                              "example": "fc05d545-6def-459d-8bbc-cf630ea98fb6"
                            },
                            "gender": {
                              "type": "string",
                              "example": "male"
                            },
                            "email": {
                              "type": "string",
                              "example": "developerteam12345@gmail.com"
                            },
                            "email_verified": {
                              "type": "string",
                              "example": "true"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "userid": {
                  "type": "number",
                  "example": "5"
                },
                "role": {
                  "type": "string",
                  "example": "user"
                }
              }
            }
          }
        }
      }
    },
    "/auth/updateProfile": {
      "post": {
        "tags": [
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/notification_server_response"
          },
          "400": {
            "$ref": "#/responses/default_server_error"
          },
          "401": {
            "$ref": "#/responses/server_authorization_error"
          }
        }
      }
    },
    "/auth/forgotPassword": {
      "post": {
        "tags": [
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/forgetPassword"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/notification_server_response"
          },
          "400": {
            "$ref": "#/responses/default_server_error"
          },
          "401": {
            "$ref": "#/responses/server_authorization_error"
          }
        }
      }
    },
    "/auth/confirmPassword": {
      "post": {
        "tags": [
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/confirmPassword"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/notification_server_response"
          },
          "400": {
            "$ref": "#/responses/default_server_error"
          },
          "401": {
            "$ref": "#/responses/server_authorization_error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "server_error_model": {
      "type": "object",
      "required": "message",
      "properties": {
        "error": {"type": "string"}
      }
    },
    "server_authorization_model":{
      "type": "object",
      "required": "message",
      "properties": {
        "error": {"type": "string"}
      }
    },
    "server_status_response" :{
      "properties":{
        "details": {
          "properties": {
            "status": "string",
            "data": {
              "type": "object"
            }
          }
        }
      },
      "type": "object"
    },
    "register": {
      "required": [
        "email",
        "familyName",
        "gender",
        "custom:role",
        "password",
        "phoneNo",
        "address"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "familyName": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "custom:role": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phoneNo": {
          "type": "number"
        },
        "address": {
          "type": "string"
        }
      }
    },
...
    "login": {
      "required": [
        "email,password"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "updateProfile": {
      "required": [
        "email",
        "familyName",
        "gender",
        "custom:role",
        "phoneNo",
        "address"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "familyName": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "custom:role": {
          "type": "string"
        },
        "phoneNo": {
          "type": "number"
        },
        "address": {
          "type": "string"
        }
      }
    },
    "forgetPassword": {
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "confirmPassword": {
      "required": [
        "code,password"
      ],
      "properties": {
        "code": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  },
  "responses":{
    "default_server_error":{
      "description": "unexpected error",
      "schema":{ "$ref": "#/definitions/server_error_model"}
    },
    "notification_server_response": {
      "description": "The server status will be returned",
      "schema":{ "$ref": "#/definitions/server_status_response"}
    },
    "server_authorization_error":{
      "description": "Authorization Error. Thrown if the user does not have access to resource.",
      "schema":{ "$ref": "#/definitions/server_authorization_model"}
    }
  },
  "securityDefinitions": {
    "": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}